Losses worked on
