+++ lib/python/layers/compilation/metrics.py
@@ -1,0 +1,10 @@
+from python.directed_acyclic_graph import Metric, NamedLayerSettings
+
+class MetricSettings(NamedLayerSettings):
+    pass
+class Accuracy(Metric):
+    pass
+class Poisson(Metric):
+    pass
+class LogCoshError(Metric):
+    pass
