+++ lib/python/layers/compilation/optimizer.py
@@ -1,0 +1,10 @@
+from python.directed_acyclic_graph import NamedLayerSettings, Optimizer
+
+class OptimizerSettings(NamedLayerSettings):
+    pass
+class RMSProp(Optimizer):
+    pass
+class Adagrad(Optimizer):
+    pass
+class Ftrl(Optimizer):
+    pass
